Added new class for json serilaisation options.
