+++ src/NoFrixion.MoneyMoov/JsonConverters/JsonStringEnumMemberConverter.cs
@@ -11,7 +11,7 @@
 // 09 Jun 2024  Aaron Clauson   Created, Stillorgan Wood, Dublin, Ireland.
 //
 // License: 
-// Proprietary NoFrixion.
+// MIT.
 //-----------------------------------------------------------------------------
 
 using System.Reflection;
